[ADD] Tabela de Historico Finalizada

+++ api/database/migrations/2019_10_26_183943_create_contas_table.php
@@ -15,7 +15,7 @@ class CreateContasTable extends Migration
     {
         Schema::create('contas', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('saldo');
+            $table->bigInteger('saldo');
             $table->boolean('conta_block');
             $table->integer('cod_operacao')->nullable();
             $table->timestamps();
